requirements.txt
grpcio
grpcio-tools
pandas
langchain
faiss-gpu
openai
tiktoken